Fix integration test bugs.
- Parallel test break read all test.

+++ test/app/outers/interfaces/deliveries/controllers/file_document_controller_test.py
@@ -46,6 +46,9 @@ async def run_around(request: pytest.FixtureRequest):
             continue
         await file_document_repository.delete_one_by_id(file_document.id)
     for document in file_document_mock_data.document_mock_data.data:
+        if request.node.name == "test__delete_one_by_id__should_delete_one_file_document__success" \
+                and document.id == file_document_mock_data.document_mock_data.data[0].id:
+            continue
         await document_repository.delete_one_by_id(document.id)
     for document_type in file_document_mock_data.document_mock_data.document_type_mock_data.data:
         await document_type_repository.delete_one_by_id(document_type.id)
@@ -122,7 +125,7 @@ async def test__patch_one_by_id__should_patch_one_file_document__success():
         file_bytes=b"file_bytes_3 patched",
     )
     response = await test_client.patch(
-        url=f"api/v1/documents/files/{file_document_mock_data.data[0].id}",
+        url=f"api/v1/documents/files/{file_document_mock_data.response_data[0].id}",
         json=json.loads(body.json())
     )
     assert response.status_code == 200
@@ -144,8 +147,8 @@ async def test__patch_one_by_id__should_patch_one_file_document__success():
 @pytest.mark.asyncio
 async def test__delete_one_by_id__should_delete_one_file_document__success():
     response = await test_client.delete(
-        url=f"api/v1/documents/files/{file_document_mock_data.data[0].id}"
+        url=f"api/v1/documents/files/{file_document_mock_data.response_data[0].id}"
     )
     assert response.status_code == 200
     content: Content[FileDocumentResponse] = Content[FileDocumentResponse](**response.json())
-    assert content.data == file_document_mock_data.data[0]
+    assert content.data == file_document_mock_data.response_data[0]
